LayoutTests/svg/dom/path-pointAtLength.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description("This tests getPointAtLength of SVG path.");

var pathElement = document.createElementNS("http://www.w3.org/2000/svg", "path");

function pointAtLengthOfPath(string) {
    pathElement.setAttributeNS(null, "d", string);

    var point = pathElement.getPointAtLength(700);
    return "(" + Math.round(point.x) + ", " + Math.round(point.y) + ")";
}

shouldThrow("pointAtLengthOfPath('M0,20 L400,20 L640,20')");
shouldThrow("pointAtLengthOfPath('M0,20 L400,20 L640,20 z')");
shouldThrow("pointAtLengthOfPath('M0,20 L400,20 z M 320,20 L640,20')");
shouldThrow("pointAtLengthOfPath('M0,20 L20,40')");
shouldThrow("pathElement.getPointAtLength(Math.NaN)");
shouldThrow("pathElement.getPointAtLength()");
shouldThrow("pathElement.getPointAtLength(Math.Infinity)");
shouldThrow("pathElement.getPointAtLength('abc')");

var successfullyParsed = true;
</script>
</body>
</html>
